requirements.txt
arxiv
paperscraper
yfinance
gensim==4.3.2
streamlit
graphviz
pandas
matplotlib-venn
pymysql
sqlalchemy
tqdm
scikit-learn
nltk
numpy
statsmodels
matplotlib
wordcloud
plotly
pathlib
datetime
orbit
orbit-ml
python-dotenv